+++ yggdrasil/command_line.py
@@ -1047,8 +1047,9 @@ class update_config(SubCommand):
         for x_true, x_false in cls.opposite_arguments:
             if getattr(args, x_false, None) is not None:
                 assert(getattr(args, x_true, None) is None)
-            setattr(args, x_true, not getattr(args, x_false))
-            delattr(args, x_false)
+                setattr(args, x_true, not getattr(args, x_false))
+            if hasattr(args, x_false):
+                delattr(args, x_false)
         lang_kwargs = {}
         for k, v in cls.language_arguments.items():
             for v_args in v: